refactor: flatten semantic assembly and parser recursion modules
Merged `src/semantic/assembly/mod.rs` and `src/semantic/assembly/model.rs` into `src/semantic/assembly.rs` to flatten the module structure.
Merged `src/parser/recursion.rs` into `src/parser/mod.rs` to inline single-function helpers.
Deleted `src/semantic/assembly/` and `src/parser/recursion.rs`.
Updated `.jules/razor.md` with the reductions.

Tests passed:
- `cargo fmt`
- `cargo clippy`
- `cargo test`
